+++ webapp/packages/supersonic-fe/src/app.tsx
@@ -131,7 +131,7 @@ export const layout: RunTimeLayoutConfig = (params) => {
           style={{ display: 'inline-block', marginTop: 8 }}
         />
         <div className="logo" style={{ position: 'relative', top: '-2px' }}>
-          SuperSonic
+          Agentic BI
         </div>
       </Space>
     ),